Move DoubleLabeledGraph from lib to server
